requirements.txt
# YouTube Automation Pipeline - Dependencies

# Anthropic Claude API
anthropic>=0.30.0

# Google Cloud TTS
google-cloud-texttospeech>=2.16.0

# YouTube Data API + Analytics API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# HTTP requests (Pexels, Pixabay)
requests>=2.31.0

# Audio processing (BGM mixing)
pydub>=0.25.1

# Config management
pyyaml>=6.0.1

# Local development (not needed in CI)
python-dotenv>=1.0.0
